evapi: Removed cast of bind port to short, as this restricted ports to signed 16-bit int values (< 32768).
htons() takes unsigned 16-bit integer and returns the same. When this value was cast to (short), higher ports could not be used.

(cherry picked from commit e9545ae1e490ff768270d03ea28f14b751d2c132)

+++ src/modules/evapi/evapi_mod.c
@@ -148,7 +148,7 @@ static int mod_init(void)
 		p = strchr(_evapi_bind_param, ':');
 		if(p!=NULL) {
 			*p++ = '\0';
-			_evapi_bind_port = (short)atoi(p);
+			_evapi_bind_port = atoi(p);
 			if (_evapi_bind_port <= 0) {
 				LM_ERR("invalid port: %d\n", _evapi_bind_port);
 				return -1;
